Modulize user management.
