+++ gemp-module/gemp-stccg-client/src/main/web/includes/admin/userAdmin.html
@@ -1,7 +1,3 @@
-
-<script type="text/javascript" src="js/gemp-022/manage.js"></script>
-
-
 <div id="banMain" class="article">
 	<h1>Reset User Password</h1>
 	<div>
